requirements.txt
pandas
selenium
webdriver-manager
sqlalchemy
psycopg2-binary
python-dotenv
numpy
